terraform_project/modules/etl/connector/app/requirements.txt
requests
flask<3.0,>=1.0
google-cloud-bigquery
aiohttp
